+++ src/util.hpp
@@ -3,11 +3,11 @@
 
 namespace TradingEngine::Util {
     struct ExchangeTime {
-        uint8_t hours;
-        uint8_t minutes;
-        uint8_t seconds;
-        uint32_t millis;
-        uint32_t micros;
+        unsigned int hours;
+        unsigned int minutes;
+        unsigned int seconds;
+        unsigned int millis;
+        unsigned int micros;
 
         ExchangeTime();
         std::string getString();
